Use processor.
