+++ src/Buildalyzer/Logging/EventProcessor.cs
@@ -1,11 +1,22 @@
 extern alias StructuredLogger;
+
+using System.IO;
+using Buildalyzer.Processors;
 using Microsoft.Build.Framework;
+using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
 
 namespace Buildalyzer.Logging;
 
 internal class EventProcessor : IDisposable
 {
+    private static readonly List<CommandLineProcessor> Processors =
+    [
+        new CSharpCommandLineProcessor(),
+        new FSharpCommandLineProcessor(),
+        new VisualBasicCommandLineProcessor(),
+    ];
+
     private readonly Dictionary<string, AnalyzerResult> _results = new Dictionary<string, AnalyzerResult>();
     private readonly Stack<AnalyzerResult> _currentResult = new Stack<AnalyzerResult>();
     private readonly Stack<TargetStartedEventArgs> _targetStack = new Stack<TargetStartedEventArgs>();
@@ -156,27 +167,12 @@ internal class EventProcessor : IDisposable
             return;
         }
 
-        // Process the command line arguments for the Fsc task
-        if (e.SenderName?.Equals("Fsc", StringComparison.OrdinalIgnoreCase) == true)
+        if (Processors.Find(p => p.IsApplicable(e)) is { } processor)
         {
-            result.ProcessFscCommandLine(e.Message);
-        }
-
-        // Process the command line arguments for the Csc task
-        if (e is TaskCommandLineEventArgs cmd
-            && string.Equals(cmd.TaskName, "Csc", StringComparison.OrdinalIgnoreCase))
-        {
-            result.ProcessCscCommandLine(cmd.CommandLine);
-        }
-
-        if (e is TaskCommandLineEventArgs cmdVbc &&
-            string.Equals(cmdVbc.TaskName, "Vbc", StringComparison.OrdinalIgnoreCase))
-        {
-            result.ProcessVbcCommandLine(cmdVbc.CommandLine);
+            result.SetCompilerCommand(processor.Parse(e.Message, new FileInfo(result.ProjectFilePath).Directory));
         }
 
         bool IsRelevant() => string.IsNullOrEmpty(result.Command) || AnalyzerManager.NormalizePath(e.ProjectFile) == _projectFilePath;
-
     }
 
     private void BuildFinished(object sender, BuildFinishedEventArgs e)
